+++ DomainManagement/functions/gppermissions/Register-DMGPPermission.ps1
@@ -70,6 +70,7 @@
 
 		Reads all settings from the provided json file and registers them.
 	#>
+	[Diagnostics.CodeAnalysis.SuppressMessageAttribute("PSReviewUnusedParameter", "")]
 	[CmdletBinding()]
 	param (
 		[Parameter(Mandatory = $true, ValueFromPipelineByPropertyName = $true, ParameterSetName = 'Explicit')]