+++ DomainManagement/internal/functions/acl_ace/Remove-RedundantAce.ps1
@@ -23,6 +23,7 @@
 
         Removes all redundant access rules on $aclobject that apply to $identity.
     #>
+	[Diagnostics.CodeAnalysis.SuppressMessageAttribute("PSReviewUnusedParameter", "")]
     [Diagnostics.CodeAnalysis.SuppressMessageAttribute("PSUseShouldProcessForStateChangingFunctions", "")]
     [CmdletBinding()]
     param (